Fix up browser full-screen aspect ratio
